+++ src/components/editor.tsx
@@ -4,12 +4,6 @@ import StartModal from "@/components/modal/startModal";
 import ListView from "@/components/waypointList/ListView";
 import dynamic from "next/dynamic";
 import Editor from "@/components/waypointEditor/editor";
-import { useMap } from "@/util/context/MapContext";
-import { useWaypoints } from "@/util/context/WaypointContext";
-import { avgLatLng } from "@/lib/world/distance";
-import { filterLatLngAltCmds } from "@/lib/commands/commands";
-import { getLatLng } from "@/util/WPCollection";
-import { useEffect } from "react";
 
 export function WMEditor() {
