+++ src/components/map/WaypointMap.tsx
@@ -6,14 +6,14 @@ import { useWaypoints } from "../../util/context/WaypointContext";
 import { Tool } from "@/types/tools";
 import { LeafletMouseEvent, Map } from "leaflet";
 import { useEffect, useRef } from "react";
-import { getLatLng, MoveWPsAvgTo } from "@/util/WPCollection";
 import { defaultDoLandStart, defaultTakeoff, defaultWaypoint } from "@/lib/mission/defaults";
 import { useMap } from '@/util/context/MapContext';
 import MapController from "./mapController";
 import MapLayers from "./layers/layers";
 import { makeCommand } from "@/lib/commands/default";
-import { avgLatLng } from "@/lib/world/distance";
-import { filterLatLngAltCmds } from "@/lib/commands/commands";
+import { Command, filterLatLngAltCmds, filterLatLngCmds } from "@/lib/commands/commands";
+import { Node } from "@/types/waypoints";
+import { avgLatLng, getLatLng } from "@/lib/world/latlng";
 
 export default function MapStuff() {
   const { waypoints, setWaypoints, activeMission, tool, setTool, selectedWPs } = useWaypoints()
@@ -111,8 +111,34 @@ export default function MapStuff() {
       case "Place": {
         setTool("Waypoint")
         setWaypoints((waypoints) => {
-          MoveWPsAvgTo(e.latlng, waypoints, selectedWPs, activeMission)
-          return waypoints
+          const mission = waypoints.get(activeMission);
+
+          let wps: Node[] = [];
+          let wpsIds: number[] = [];
+          if (selectedWPs.length === 0) {
+            wps = mission;
+            wpsIds = mission.map((_, index) => index);
+          } else {
+            wps = mission.filter((_, id) => selectedWPs.includes(id));
+            wpsIds = selectedWPs;
+          }
+
+          const leaves = wps.map((x) => waypoints.flattenNode(x)).reduce((cur, acc) => (acc.concat(cur)), [])
+
+          const avgll = avgLatLng(filterLatLngCmds(leaves).map(getLatLng))
+          if (avgll == undefined) { return waypoints }
+          const { lat, lng } = avgll
+          let waypointsUpdated = waypoints.clone();
+          for (let i = 0; i < wps.length; i++) {
+            waypointsUpdated.changeParam(wpsIds[i], activeMission, (cmd: Command) => {
+              if ("latitude" in cmd.params && "longitude" in cmd.params) {
+                cmd.params.latitude += e.latlng.lat - lat
+                cmd.params.longitude += e.latlng.lng - lng
+              }
+              return cmd;
+            });
+          }
+          return waypointsUpdated;
         })
         break;
       }
