+++ src/util/context/VehicleTypeProvider.tsx
@@ -6,12 +6,13 @@ import { Vehicle } from '@/lib/vehicles/types';
 
 type Props = {
   children: React.ReactNode;
+  vehicle?: Vehicle
 };
 
-export default function VehicleProvider({ children }: Props) {
-  const [vehicle, setVehicle] = useState(defaultPlane as Vehicle)
+export default function VehicleProvider({ children, vehicle }: Props) {
+  const [v, setVehicle] = useState(vehicle || defaultPlane as Vehicle)
   return (
-    <vehicleTypeContext.Provider value={{ vehicle: vehicle, setVehicle: setVehicle }}>
+    <vehicleTypeContext.Provider value={{ vehicle: v, setVehicle: setVehicle }}>
       {children}
     </vehicleTypeContext.Provider>
   );
